refactor(providers): break the registry/attribution import cycle
Extract resolveRoutingIdentity and ConnectionResolutionError out of
connection-resolution.ts into a new leaf module,
providers/routing-identity.ts, and point usage/attribution.ts at it.

attribution.ts importing resolveRoutingIdentity from
connection-resolution.ts was the back-edge of the module graph's only
import cycle:

  connection-resolution -> inference/connections -> registry ->
  inference/adapter-factory -> retry -> usage/attribution ->
  connection-resolution

That cycle is what forced the compiled binary's cyclic-module lowering
onto the plugin-api barrel graph and produced the TDZ ReferenceError
that #38835 worked around at the facade level. With this change the
barrel's transitive import graph (286 modules) contains zero cycles.

The new module's transitive closure is 6 leaf modules (codecs,
constants, error base classes), none of which reach the registry.
connection-resolution.ts re-exports both symbols so its existing
importers are untouched; only attribution.ts must import from the new
path, since going through the re-export would re-form the cycle.
